Replace StepSubType test with updated assertions
Remove the old StepSubType test and add a new StepSubTypeTest that reflects the current enum labels. The new tests verify that toString() matches getSubType() for every enum constant and include explicit assertions for each known subtype string (information messages, step completion types, GUI actions, cell step variants, hints, default, etc.). Removed previous null/blank checks and the valueOf-based lookup test in favor of the consolidated, explicit expectations.
